Move utils unit tests to separate file
