+++ integration/test_filesystems_integration.py
@@ -22,7 +22,6 @@ import aiohttp
 import pandas as pd
 import pytest
 import v3io
-from fsspec.implementations.local import LocalFileSystem
 
 from integration.integration_test_utils import V3ioHeaders, _generate_table_name
 from storey import (
@@ -39,7 +38,6 @@ from storey import (
     build_flow,
 )
 from storey.dtypes import V3ioError
-from storey.utils import get_remaining_path, url_to_file_system
 
 
 @pytest.fixture()
@@ -666,17 +664,3 @@ def test_filter_before_after_partitioned_outer_other_partition(setup_teardown_te
     ]
 
     assert read_back_result == expected, f"{read_back_result}\n!=\n{expected}"
-
-
-def test_get_path_utils():
-    url = "wasbs://mycontainer@myaccount.blob.core.windows.net/path/to/object.csv"
-    schema, path = get_remaining_path(url)
-    assert path == "mycontainer/path/to/object.csv"
-    assert schema == "wasbs"
-
-
-def test_ds_get_path_utils():
-    url = "ds://:file@profile/path/to/object.csv"
-    fs, path = url_to_file_system(url, "")
-    assert path == "/path/to/object.csv"
-    assert isinstance(fs, LocalFileSystem)
